Updated ScrollTop component to TypeScript and confirmed TypeScript is now working in app
