+++ src/pages/components-overview/AntIcons.js
@@ -2,6 +2,7 @@
 import { styled } from '@mui/material/styles';
 
 // project import
+// import { ComponentSkeleton } from './ComponentSkeleton';
 import ComponentSkeleton from './ComponentSkeleton';
 import MainCard from 'components/MainCard';
